influx/requirements.txt
influxdb
kafka-python
pytz
requests
websocket-client
